feat: add validation stock availability before deleting it

+++ modules/stock/stock.service.ts
@@ -1,3 +1,4 @@
+import { AppError } from "@/lib/errors";
 import { stockRepository } from "./stock.repository";
 import { InsertStock } from "./stock.types";
 
@@ -14,7 +15,21 @@ export const stockService = {
     return stockRepository.updateStock(id, data);
   },
 
-  deleteStock(id: number) {
+  async deleteStock(id: number) {
+    const stock = await stockRepository.getStockById(id);
+
+    if (!stock) {
+      throw new AppError("Data stok tidak ditemukan", 404);
+    }
+
+    if (stock.initial_stock !== 0 || stock.ending_stock !== 0) {
+      throw new AppError("Stok awal dan stok akhir harus 0", 400);
+    }
+
+    if (stock.qty_in !== 0 || stock.qty_out !== 0) {
+      throw new AppError("Masih ada transaksi dengan barang ini", 400);
+    }
+
     return stockRepository.deleteStock(id);
   },
 };
